Combine separate TruthSeed modules into a single module with templated base class.
